New tiles, enemy health bars, mouse sensitivity changes
New tiles have been created to display the functionality of the tile generator.
Enemy health bars have been implemented. They are billboards that always face the player. Having this enabled at all times was bad for framerate so this only works if the player is within a certain radius.
Mouse sensitivity is calculated slightly different, namely it no longer is dependent on frames. This means it no longer uses the Time.deltaTime constant.