Implemented Menu System
Main menu and in game pause menu have both been implemented.
The main menu has an options menu which for now has a silder and input box for volume and sensitivity but the actual functionality is not implemented yet.
This also goes for the Options button in the pause menu, it exists but serves no function.